+++ fhir/resources/DSTU2/__init__.py
@@ -1,0 +1,6 @@
+"""FHIR (version 1.0.2.7202) DSTU2 resources.
+This whole package is generated using
+https://github.com/smart-on-fhir/fhir-parser/tree/b9140c8a189b9cf7efe882760aa55ad88dc361ef
+
+All credit: SMART Health IT.
+"""