+++ fhir/resources/DSTU2/duration.py
@@ -1,0 +1,29 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+#
+#  Generated from FHIR 1.0.2.7202 (http://hl7.org/fhir/StructureDefinition/Duration) on 2019-05-14.
+#  2019, SMART Health IT.
+
+
+from . import quantity
+
+class Duration(quantity.Quantity):
+    """ A length of time.
+    
+    There SHALL be a code if there is a value and it SHALL be an expression of
+    time.  If system is present, it SHALL be UCUM.
+    """
+    
+    resource_name = "Duration"
+    
+    def __init__(self, jsondict=None, strict=True):
+        """ Initialize all valid properties.
+        
+        :raises: FHIRValidationError on validation errors, unless strict is False
+        :param dict jsondict: A JSON dictionary to use for initialization
+        :param bool strict: If True (the default), invalid variables will raise a TypeError
+        """
+        
+        super(Duration, self).__init__(jsondict=jsondict, strict=strict)
+
+