+++ fhir/resources/DSTU2/relatedperson.py
@@ -1,0 +1,96 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+#
+#  Generated from FHIR 1.0.2.7202 (http://hl7.org/fhir/StructureDefinition/RelatedPerson) on 2019-05-14.
+#  2019, SMART Health IT.
+
+
+from . import domainresource
+
+class RelatedPerson(domainresource.DomainResource):
+    """ An person that is related to a patient, but who is not a direct target of
+    care.
+    
+    Information about a person that is involved in the care for a patient, but
+    who is not the target of healthcare, nor has a formal responsibility in the
+    care process.
+    """
+    
+    resource_name = "RelatedPerson"
+    
+    def __init__(self, jsondict=None, strict=True):
+        """ Initialize all valid properties.
+        
+        :raises: FHIRValidationError on validation errors, unless strict is False
+        :param dict jsondict: A JSON dictionary to use for initialization
+        :param bool strict: If True (the default), invalid variables will raise a TypeError
+        """
+        
+        self.address = None
+        """ Address where the related person can be contacted or visited.
+        List of `Address` items (represented as `dict` in JSON). """
+        
+        self.birthDate = None
+        """ The date on which the related person was born.
+        Type `FHIRDate` (represented as `str` in JSON). """
+        
+        self.gender = None
+        """ male | female | other | unknown.
+        Type `str`. """
+        
+        self.identifier = None
+        """ A human identifier for this person.
+        List of `Identifier` items (represented as `dict` in JSON). """
+        
+        self.name = None
+        """ A name associated with the person.
+        Type `HumanName` (represented as `dict` in JSON). """
+        
+        self.patient = None
+        """ The patient this person is related to.
+        Type `FHIRReference` referencing `Patient` (represented as `dict` in JSON). """
+        
+        self.period = None
+        """ Period of time that this relationship is considered valid.
+        Type `Period` (represented as `dict` in JSON). """
+        
+        self.photo = None
+        """ Image of the person.
+        List of `Attachment` items (represented as `dict` in JSON). """
+        
+        self.relationship = None
+        """ The nature of the relationship.
+        Type `CodeableConcept` (represented as `dict` in JSON). """
+        
+        self.telecom = None
+        """ A contact detail for the person.
+        List of `ContactPoint` items (represented as `dict` in JSON). """
+        
+        super(RelatedPerson, self).__init__(jsondict=jsondict, strict=strict)
+    
+    def elementProperties(self):
+        js = super(RelatedPerson, self).elementProperties()
+        js.extend([
+            ("address", "address", address.Address, True, None, False),
+            ("birthDate", "birthDate", fhirdate.FHIRDate, False, None, False),
+            ("gender", "gender", str, False, None, False),
+            ("identifier", "identifier", identifier.Identifier, True, None, False),
+            ("name", "name", humanname.HumanName, False, None, False),
+            ("patient", "patient", fhirreference.FHIRReference, False, None, True),
+            ("period", "period", period.Period, False, None, False),
+            ("photo", "photo", attachment.Attachment, True, None, False),
+            ("relationship", "relationship", codeableconcept.CodeableConcept, False, None, False),
+            ("telecom", "telecom", contactpoint.ContactPoint, True, None, False),
+        ])
+        return js
+
+
+from . import address
+from . import attachment
+from . import codeableconcept
+from . import contactpoint
+from . import fhirdate
+from . import fhirreference
+from . import humanname
+from . import identifier
+from . import period