+++ fhir/resources/DSTU2/sampleddata.py
@@ -1,0 +1,71 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+#
+#  Generated from FHIR 1.0.2.7202 (http://hl7.org/fhir/StructureDefinition/SampledData) on 2019-05-14.
+#  2019, SMART Health IT.
+
+
+from . import element
+
+class SampledData(element.Element):
+    """ A series of measurements taken by a device.
+    
+    A series of measurements taken by a device, with upper and lower limits.
+    There may be more than one dimension in the data.
+    """
+    
+    resource_name = "SampledData"
+    
+    def __init__(self, jsondict=None, strict=True):
+        """ Initialize all valid properties.
+        
+        :raises: FHIRValidationError on validation errors, unless strict is False
+        :param dict jsondict: A JSON dictionary to use for initialization
+        :param bool strict: If True (the default), invalid variables will raise a TypeError
+        """
+        
+        self.data = None
+        """ Decimal values with spaces, or "E" | "U" | "L".
+        Type `str`. """
+        
+        self.dimensions = None
+        """ Number of sample points at each time point.
+        Type `int`. """
+        
+        self.factor = None
+        """ Multiply data by this before adding to origin.
+        Type `float`. """
+        
+        self.lowerLimit = None
+        """ Lower limit of detection.
+        Type `float`. """
+        
+        self.origin = None
+        """ Zero value and units.
+        Type `Quantity` referencing `SimpleQuantity` (represented as `dict` in JSON). """
+        
+        self.period = None
+        """ Number of milliseconds between samples.
+        Type `float`. """
+        
+        self.upperLimit = None
+        """ Upper limit of detection.
+        Type `float`. """
+        
+        super(SampledData, self).__init__(jsondict=jsondict, strict=strict)
+    
+    def elementProperties(self):
+        js = super(SampledData, self).elementProperties()
+        js.extend([
+            ("data", "data", str, False, None, True),
+            ("dimensions", "dimensions", int, False, None, True),
+            ("factor", "factor", float, False, None, False),
+            ("lowerLimit", "lowerLimit", float, False, None, False),
+            ("origin", "origin", quantity.Quantity, False, None, True),
+            ("period", "period", float, False, None, True),
+            ("upperLimit", "upperLimit", float, False, None, False),
+        ])
+        return js
+
+
+from . import quantity